+++ lib/Cellular.php
@@ -19,7 +19,16 @@ namespace DTL\Cellular;
 abstract class Cellular extends Collection implements CellularInterface
 {
     /**
-     * {@inheritDoc}
+     * @var array
+     */
+    protected $groups = array();
+
+    /**
+     * Return the values of all the cells contained in this cellular instance.
+     *
+     * @param array $groups
+     *
+     * @return array
      */
     public function getValues(array $groups = array())
     {
@@ -32,7 +41,36 @@ abstract class Cellular extends Collection implements CellularInterface
     }
 
     /**
-     * {@inheritDoc}
+     * Return all of the groups to which this cellular instance belongs.
+     *
+     * @return array
+     */
+    public function getGroups()
+    {
+        return $this->groups;
+    }
+
+    /**
+     * Set the groups to which this cellular instance belongs.
+     *
+     * @param array $groups
+     *
+     * @return this
+     */
+    public function setGroups($groups)
+    {
+        $this->groups = $groups;
+
+        return $this;
+    }
+
+    /**
+     * Use a closure to assign a value to each cell in this cellular instance.
+     *
+     * @param \Closure
+     * @param array $groups
+     *
+     * @return this
      */
     public function mapValues(\Closure $closure, array $groups = array())
     {
@@ -40,6 +78,8 @@ abstract class Cellular extends Collection implements CellularInterface
             $value = $closure($cell);
             $cell->setValue($value);
         }
+
+        return $this;
     }
 
     abstract public function getCells(array $groups = array());